Refactor: 메인 페이지를 /home에서 루트(/)로 변경하여 SEO 색인 문제 해결
- /home 경로 제거, 루트(/)에서 바로 콘텐츠 제공하여 리다이렉트 제거
- 기존 /home, /ko/home URL은 301 영구 리다이렉트로 루트로 통합
- canonical, alternates, sitemap URL을 /, /en으로 업데이트
- 병렬 라우트(@chat, @side)를 app/[locale]/로 이동
