+++ app/[locale]/@chat/_components/question.tsx
