+++ app/[locale]/@chat/_lib/constants.ts
