+++ app/[locale]/@chat/body.tsx
