+++ app/[locale]/@chat/page.tsx
