+++ app/[locale]/@side/_components/profileSwiper.tsx
