+++ app/[locale]/@side/page.tsx
